refactor(orchestrator): extract ToolApprovalGateSupport
R2 step 4 — per-turn pause accounting, the approver-facing pause reason, the
durable PendingToolCallBatch snapshot and its size caps, the batch
fingerprint, tool-call-id normalisation, and the pause-cap guard's
metric/audit emission. Pure move, no behavior change.

AgentOrchestrator: 2,127 -> 1,904 lines, finally under the 2,000-line
Checkstyle FileLength limit it has exceeded for its entire history.

The cluster accompanies ToolApprovalGate: that class decides whether a batch
pauses, everything here is what happens once it does. Almost all of it is
static, which is why the move was mechanical.

All of it stays as declared delegators on AgentOrchestrator. Bare-token sweep
across the test tree first: fingerprint 25 refs, maxPausesPerTurn 16,
buildPauseReason 13, normalizeToolCallIds 10, buildPendingBatch 10,
readToolPauseCount 3. Four are reached via getDeclaredMethod, which resolves
only methods declared on that exact class, and buildPendingBatch is called
directly as an instance method eight times. Signatures and modifiers
preserved verbatim.

442 tests green across the orchestrator, gate and HITL batteries.
